fix(formathtmlpreview): moved helper function to stories folder and updated all imports in stories
