+++ packages/core/src/components/radio-button/radio-button.stories.tsx
@@ -1,4 +1,4 @@
-import { formatHtmlPreview } from '../../utils/utils';
+import formatHtmlPreview from '../../stories/formatHtmlPreview';
 import readme from './readme.md';
 import { ComponentsFolder } from '../../utils/constants';
