+++ packages/core/src/components/tabs/navigation-tabs/navigation-tabs.stories.tsx
@@ -1,4 +1,4 @@
-import { formatHtmlPreview } from '../../../utils/utils';
+import formatHtmlPreview from '../../../stories/formatHtmlPreview';
 import readme from './readme.md';
 import readmeTab from './navigation-tab/readme.md';
 import { ComponentsFolder } from '../../../utils/constants';
