+++ packages/core/src/components/text-field/text-field.stories.tsx
@@ -1,5 +1,5 @@
 import readme from './readme.md';
-import { formatHtmlPreview } from '../../utils/utils';
+import formatHtmlPreview from '../../stories/formatHtmlPreview';
 import { ComponentsFolder } from '../../utils/constants';
 
 export default {
