+++ packages/core/src/stories/patterns/navigation/navigation-fewitems.stories.tsx
@@ -1,4 +1,5 @@
-import { formatHtmlPreview } from '../../../utils/utils';
+import formatHtmlPreview from '../../formatHtmlPreview';
+
 import readme from './readme.md';
 
 export default {
