+++ packages/core/src/stories/utility/color/text-color.stories.tsx
@@ -1,4 +1,4 @@
-import { formatHtmlPreview } from '../../../utils/utils';
+import formatHtmlPreview from '../../formatHtmlPreview';
 
 export default {
   title: 'Utilities/Colors',
